starting to implement some tighter unit testing patterns
